utils/requirements.txt
pyyaml
opml

feedparser
requests
feedgen
pytz

BeautifulSoup